requirements.txt
# VCD + AGLA Combined Method - Python Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.31.0

# LLaVA dependencies
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
einops>=0.6.1
einops-exts>=0.0.4
timm>=0.9.0

# AGLA dependencies (BLIP-ITM)
salesforce-lavis>=1.0.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Evaluation
tqdm>=4.65.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Logging
tensorboard>=2.13.0

# Optional: For better performance
ninja>=1.11.0
flash-attn>=2.0.0  # Optional, for faster attention

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Note: You may need to install some dependencies from source or specific versions
# depending on your CUDA version and system configuration.

# Installation instructions:
# 1. Install PyTorch with CUDA support:
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 2. Install transformers and accelerate:
#    pip install transformers accelerate
#
# 3. Install LAVIS for AGLA:
#    pip install salesforce-lavis
#
# 4. Install remaining dependencies:
#    pip install -r requirements.txt
#
# 5. Install LLaVA (if not already installed):
#    You may need to clone and install from the AGLA or VCD project's llava directory

